+++ CHANGES.md
@@ -2,6 +2,27 @@
 
 ## Unreleased
 
+## 2.5
+
+## Indent run keywords
+
+New ``IndentNestedKeywords`` transformer for formatting whitespace in run keywords variants such as ``Run Keywords`` or 
+``Run Keyword And Continue On Failure``. It will result in transforming following code:
+```robotframework
+Run Keyword    Run Keyword If    ${True}    Run keywords   Log    foo    AND    Log    bar    ELSE    Log    baz
+```
+to:
+```robotframework
+Run Keyword
+...    Run Keyword If    ${True}
+...        Run keywords
+...            Log    foo
+...            AND
+...            Log    bar
+...    ELSE
+...        Log    baz
+```
+
 # Remove empty lines in multiline statements
 ``NormalizeNewLines`` transformer now removes empty lines inside multiline statements:
 ```robotframework